requirements.txt
authlib
flask~=1.1.2
google-api-python-client
google-auth
virtualenv
PyMySQL~=0.9.3
numpy~=1.17.2